refactor nix-shell in activation script

+++ home-manager.nix
@@ -65,9 +65,11 @@
       ''
           # https://www.lafuente.me/posts/installing-home-manager/
           # fishPlugins.foreign-env
-          echo "export NIX_PATH=\$HOME/.nix-defexpr/channels''\${NIX_PATH:+:}\$NIX_PATH" > ./.bashrc
+          
+          if not test -e ~/.bashrc
+            echo "export NIX_PATH=\$HOME/.nix-defexpr/channels''\${NIX_PATH:+:}\$NIX_PATH" > ./.bashrc
+          end
           fenv source ./.bashrc
-          nix-shell https://github.com/rycee/home-manager/archive/release-18.03.tar.gz -A install
 
           # nix
           if test -e /nix/var/nix/profiles/default/etc/profile.d/nix-daemon.sh
